+++ src/entities/index.ts
@@ -1,2 +1,3 @@
 export * from "./Zombie";
 export * from './FallingBlock';
+export * from './ElderGuardian';